Lua!! Added Script class.
Lua now works! Next stop: the interface!

+++ lib/LuaIntf/LuaIntf/LuaState.h
@@ -585,7 +585,11 @@ public:
         { lua_len(L, table_idx); }
 
     int tableLen(int table_idx) const
-        { return luaL_len(L, table_idx); }
+        {
+			// Casting as an int to shush the warning.
+			// Hopefully this will be fixed in the future!
+			return (int)luaL_len(L, table_idx);
+		}
 
 // set functions (stack -> Lua)
